+++ src/browser.rs
@@ -3,7 +3,7 @@ pub mod service;
 pub mod url;
 pub mod util;
 
-#[cfg(any(feature = "serde-wasm-bindgen", feature = "serde-json"))]
+#[cfg(feature = "routing")]
 mod json;
 
 pub use url::{Url, UrlSearch, DUMMY_BASE_URL};